+++ jspwiki-main/src/main/java/org/apache/wiki/plugin/ListLocksPlugin.java
@@ -26,6 +26,7 @@ import org.apache.wiki.pages.PageManager;
 import org.apache.wiki.preferences.Preferences;
 
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.ResourceBundle;
 
@@ -38,6 +39,17 @@ import java.util.ResourceBundle;
  */
 public class ListLocksPlugin implements Plugin {
 
+    @Override
+    public String getDisplayName(Locale locale) {
+        final ResourceBundle rb = ResourceBundle.getBundle(PluginManager.PLUGIN_RESOURCE_LOCATION, locale);
+        return rb.getString(this.getClass().getSimpleName());
+    }
+   
+    
+    @Override
+    public  String getSnipExample() {
+        return "ListLocksPlugin";
+    }
     /**
      *  {@inheritDoc}
      */