+++ jspwiki-main/src/main/java/org/apache/wiki/plugin/RecentChangesPlugin.java
@@ -45,7 +45,9 @@ import java.util.Calendar;
 import java.util.Collection;
 import java.util.Date;
 import java.util.GregorianCalendar;
+import java.util.Locale;
 import java.util.Map;
+import java.util.ResourceBundle;
 
 
 /**
@@ -76,6 +78,13 @@ public class RecentChangesPlugin extends AbstractReferralPlugin implements Plugi
     public static final String DEFAULT_DATE_FORMAT ="dd.MM.yyyy";
 
 
+    @Override
+    public String getDisplayName(Locale locale) {
+        final ResourceBundle rb = ResourceBundle.getBundle(PluginManager.PLUGIN_RESOURCE_LOCATION, locale);
+        return rb.getString(this.getClass().getSimpleName());
+    }
+   
+    
     /**
      * {@inheritDoc}
      */