+++ jspwiki-main/src/main/java/org/apache/wiki/plugin/UnusedPagesPlugin.java
@@ -24,7 +24,9 @@ import org.apache.wiki.references.ReferenceManager;
 import org.apache.wiki.util.TextUtil;
 
 import java.util.Collection;
+import java.util.Locale;
 import java.util.Map;
+import java.util.ResourceBundle;
 
 /**
  * Plugin for displaying pages that are not linked to in other pages.
@@ -41,6 +43,12 @@ public class UnusedPagesPlugin extends AbstractReferralPlugin {
     /** If set to "true", attachments are excluded from display.  Value is {@value}. */
     public static final String PARAM_EXCLUDEATTS = "excludeattachments";
 
+    @Override
+    public String getDisplayName(Locale locale) {
+        final ResourceBundle rb = ResourceBundle.getBundle(PluginManager.PLUGIN_RESOURCE_LOCATION, locale);
+        return rb.getString(this.getClass().getSimpleName());
+    } 
+    
     /**
      *  {@inheritDoc}
      */