+++ jspwiki-main/src/main/java/org/apache/wiki/search/DefaultSearchManager.java
@@ -51,9 +51,15 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
+import org.apache.wiki.api.plugin.Plugin;
+import org.apache.wiki.i18n.InternationalizationManager;
+import org.apache.wiki.plugin.PluginManager;
+import org.apache.wiki.preferences.Preferences;
+import static org.apache.wiki.search.SearchManager.PLUGIN_SEARCH;
 
 
 /**
@@ -80,8 +86,81 @@ public class DefaultSearchManager extends BasePageFilter implements SearchManage
 
         // TODO: Replace with custom annotations. See JSPWIKI-566
         WikiAjaxDispatcherServlet.registerServlet( JSON_SEARCH, new JSONSearch() );
+        WikiAjaxDispatcherServlet.registerServlet( PLUGIN_SEARCH, new PluginSearch() );
     }
 
+    /**
+     *  Provides a JSON AJAX API to the JSPWiki Plugin discovery mechanism,
+     *  primarily used for [{}] based auto complete
+     */
+    public class PluginSearch implements WikiAjaxServlet {
+
+        public static final String AJAX_ACTION_PLUGINS = "plugins";
+        public static final int DEFAULT_MAX_RESULTS = 20;
+        public int maxResults = DEFAULT_MAX_RESULTS;
+
+        /** {@inheritDoc} */
+        @Override
+        public String getServletMapping() {
+            return PLUGIN_SEARCH;
+        }
+        public static class SimpleSnipData {
+            public String displayName;
+            public String snip;
+        }
+
+        /** {@inheritDoc} */
+        @Override
+        public void service( final HttpServletRequest req,
+                             final HttpServletResponse resp,
+                             final String actionName,
+                             final List< String > params ) throws IOException {
+            String result = "[]";
+            String itemId = null;
+            if( actionName != null && StringUtils.isNotBlank( actionName ) ) {
+                if( !params.isEmpty() ) {
+                    itemId = params.get( 0 );
+                    LOG.debug( "itemId=" + itemId );
+                    if( params.size() > 1 ) {
+                        itemId = params.get( 0 );
+                        final String maxResultsParam = params.get( 1 );
+                        LOG.debug( "maxResultsParam=" + maxResultsParam );
+                        if( StringUtils.isNotBlank( maxResultsParam ) && StringUtils.isNumeric( maxResultsParam ) ) {
+                            maxResults = Integer.parseInt( maxResultsParam );
+                        }
+                    }
+                }
+
+                if( actionName.equals( AJAX_ACTION_PLUGINS ) ) {
+                    LOG.debug( "Calling getPlugins() START" );
+                    PluginManager mgr = m_engine.getManager(PluginManager.class);
+                    final List< Plugin > plugins = mgr.getDiscoveredPlugins();
+                    List< SimpleSnipData > callResults = new ArrayList<>();
+                    final Context wikiContext = Wiki.context().create(m_engine, req, "pluginDisco");
+                    Locale locale =  Preferences.getLocale(wikiContext);
+                    for (Plugin p : plugins) {
+                        if ( itemId != null && StringUtils.isNotBlank(itemId) ) {
+                            if ( !p.getSnipExample().startsWith(itemId) ) {
+                                continue;
+                            }
+                        } else {
+                            SimpleSnipData data = new SimpleSnipData();
+                            data.snip = p.getSnipExample();
+                            data.displayName = p.getDisplayName(locale);
+                            callResults.add(data);
+                        }
+                        if ( callResults.size() > maxResults )
+                            break;
+                    }
+                    LOG.debug( "Calling getSuggestions() DONE. " + callResults.size() );
+                    result = AjaxUtil.toJson( callResults );
+                } 
+            }
+            LOG.debug( "result=" + result );
+            resp.getWriter().write( result );
+        }
+    }
+    
     /**
      *  Provides a JSON AJAX API to the JSPWiki Search Engine.
      */
@@ -160,7 +239,7 @@ public class DefaultSearchManager extends BasePageFilter implements SearchManage
                 final String cleanWikiName = MarkupParser.cleanLink(wikiName).toLowerCase() + filename;
                 final String oldStyleName = MarkupParser.wikifyLink(wikiName).toLowerCase() + filename;
                 final Set< String > allPages = m_engine.getManager( ReferenceManager.class ).findCreated();
-
+               
                 int counter = 0;
                 for( final Iterator< String > i = allPages.iterator(); i.hasNext() && counter < maxLength; ) {
                     final String p = i.next();